Working on porting 2017 to 2019. Might not be a good use of time.
WPI motor control has changed a bit since the 2017 code was written and
the new API is not especially compatible with the old one --- same rough
concepts, but new names and places. So the mapping is a little gross.

Apart from issues like that, though, I think the project would build. So
I believe that at least I finally have the build system doing my
bidding.